feat: use CDP's auto-attach mechanism
In this PR, we refactor Puppeteer to make use of CDP's
auto-attach mechanism. This allows the backend to pause
new targets and give Puppeteer a chance to configure them
properly. This fixes the flakiness related to dealing with
OOPIFs and should fix a multitude of issues related to the
network interception and navigations. If those are not fixed
completely by this PR, the PR serves a solid base for fixing them

Closes #8507, #7990
Unlocks #3667

BREAKING CHANGE: With Chromium, Puppeteer will now attach
to page/iframe targets immediately.
BREAKING CHANGE: Browser.connect requires an explicit product name
when connecting to Firefox since Firefox does not support CDP's auto-attach.